Create alias based method invocation
